test: pipeline updated assertions (#5413)
