+++ tests/ui-testing/playwright-tests/pipelines.spec.js
@@ -1,10 +1,10 @@
 import { test, expect } from "./baseFixtures";
 import logData from "../../ui-testing/cypress/fixtures/log.json";
-import { log } from "console";
+// import { log } from "console";
 import logsdata from "../../test-data/logs_data.json";
 import PipelinePage from "../pages/pipelinePage";
-import { pipeline } from "stream";
-import fs from "fs";
+// import { pipeline } from "stream";
+// import fs from "fs";
 import { v4 as uuidv4 } from "uuid";
 
 test.describe.configure({ mode: "parallel" });
@@ -25,8 +25,34 @@ async function login(page) {
     .locator('[data-cy="login-password"]')
     .fill(process.env["ZO_ROOT_USER_PASSWORD"]);
   await page.locator('[data-cy="login-sign-in"]').click();
-  //     await page.waitForTimeout(4000);
-  // await page.goto(process.env["ZO_BASE_URL"]);
+  
+}
+async function ingestion(page) {
+  const orgId = process.env["ORGNAME"];
+  const streamName = "e2e_automate";
+  const basicAuthCredentials = Buffer.from(
+    `${process.env["ZO_ROOT_USER_EMAIL"]}:${process.env["ZO_ROOT_USER_PASSWORD"]}`
+  ).toString('base64');
+
+  const headers = {
+    "Authorization": `Basic ${basicAuthCredentials}`,
+    "Content-Type": "application/json",
+  };
+  const response = await page.evaluate(async ({ url, headers, orgId, streamName, logsdata }) => {
+    const fetchResponse = await fetch(`${url}/api/${orgId}/${streamName}/_json`, {
+      method: 'POST',
+      headers: headers,
+      body: JSON.stringify(logsdata)
+    });
+    return await fetchResponse.json();
+  }, {
+    url: process.env.INGESTION_URL,
+    headers: headers,
+    orgId: orgId,
+    streamName: streamName,
+    logsdata: logsdata
+  });
+  console.log(response);
 }
 
 const selectStreamAndStreamTypeForLogs = async (page, stream) => {
@@ -40,6 +66,46 @@ const selectStreamAndStreamTypeForLogs = async (page, stream) => {
     .first()
     .click({ force: true });
 };
+async function exploreStreamAndNavigateToPipeline(page, streamName) {
+  // Navigate to the streams menu
+  await page.locator('[data-test="menu-link-\\/streams-item"]').click();
+  
+  // Search for the stream
+  await page.getByPlaceholder('Search Stream').click();
+  await page.getByPlaceholder('Search Stream').fill(streamName);
+  
+  // Click on the 'Explore' button
+  await page.getByRole('button', { name: 'Explore' }).first().click();
+  
+  // Expand the log table menu
+  await page.locator('[data-test="log-table-column-1-_timestamp"] [data-test="table-row-expand-menu"]').click();
+  
+  // Navigate to the pipeline menu
+  await page.locator('[data-test="menu-link-\\/pipeline-item"]').click();
+}
+
+async function exploreStreamAndInteractWithLogDetails(page, streamName) {
+  // Navigate to the streams menu
+  await page.locator('[data-test="menu-link-\\/streams-item"]').click();
+  
+  // Search for the stream
+  await page.getByPlaceholder('Search Stream').click();
+  await page.getByPlaceholder('Search Stream').fill(streamName);
+  
+  // Click on the 'Explore' button
+  await page.getByRole('button', { name: 'Explore' }).first().click();
+  
+  // Expand the log table menu
+  await page.locator('[data-test="log-table-column-1-_timestamp"] [data-test="table-row-expand-menu"]').click();
+  
+  // Interact with the log detail key
+  await page.locator('[data-test="log-expand-detail-key-a-text"]').click();
+  
+  // Navigate to the pipeline menu
+  await page.locator('[data-test="menu-link-\\/pipeline-item"]').click();
+}
+
+
 async function deletePipeline(page, randomPipelineName) {
   // Click the back button
   await page.locator('[data-test="add-pipeline-back-btn"]').click();
@@ -97,31 +163,7 @@ test.describe("Pipeline testcases", () => {
       "Content-Type": "application/json",
     };
 
-    // const logsdata = {}; // Fill this with your actual data
-
-    // Making a POST request using fetch API
-    const response = await page.evaluate(
-      async ({ url, headers, orgId, streamName, logsdata }) => {
-        const fetchResponse = await fetch(
-          `${url}/api/${orgId}/${streamName}/_json`,
-          {
-            method: "POST",
-            headers: headers,
-            body: JSON.stringify(logsdata),
-          }
-        );
-        return await fetchResponse.json();
-      },
-      {
-        url: process.env.INGESTION_URL,
-        headers: headers,
-        orgId: orgId,
-        streamName: streamName,
-        logsdata: logsdata,
-      }
-    );
-
-    console.log(response);
+    await ingestion(page);
     await page.goto(
       `${logData.logsUrl}?org_identifier=${process.env["ORGNAME"]}`
     );
@@ -268,6 +310,9 @@ test.describe("Pipeline testcases", () => {
     const pipelineName = `pipeline-${Math.random().toString(36).substring(7)}`;
     await pipelinePage.enterPipelineName(pipelineName);
     await pipelinePage.savePipeline();
+    await ingestion(page);
+    // Verify the data ingested in destination and verify under logs page
+    await exploreStreamAndNavigateToPipeline(page, 'destination_node');
     await pipelinePage.searchPipeline(pipelineName);
     await page.waitForTimeout(1000);
     const deletePipelineButton = page.locator(
@@ -355,13 +400,13 @@ test.describe("Pipeline testcases", () => {
     await pipelinePage.enterFunctionName(randomFunctionName);
     await page.locator(".view-lines").click();
     // Type the function text with a delay to ensure each character registers
-    await page.keyboard.type(".a=2", { delay: 100 });
+    await page.keyboard.type(".a=41", { delay: 100 });
     await page.keyboard.press("Enter");
     await page.keyboard.type(".", { delay: 100 });
     await page.getByText("Note: The function will be").click();
 
     // Check if the required text is present in the editor
-    await page.getByText(".a=2 .");
+    await page.getByText(".a=41 .");
 
     // Optional: Add a short wait to confirm the action is processed
     await page.waitForTimeout(1000);
@@ -391,6 +436,9 @@ test.describe("Pipeline testcases", () => {
     const pipelineName = `pipeline-${Math.random().toString(36).substring(7)}`;
     await pipelinePage.enterPipelineName(pipelineName);
     await pipelinePage.savePipeline();
+    await ingestion(page);
+    // Verify the data ingested in destination & function and verify under logs page 
+    await exploreStreamAndInteractWithLogDetails(page, 'destination_node');
     await pipelinePage.searchPipeline(pipelineName);
     await page.waitForTimeout(1000);
     const deletePipelineButton = page.locator(
@@ -530,6 +578,9 @@ test.describe("Pipeline testcases", () => {
     const pipelineName = `pipeline-${Math.random().toString(36).substring(7)}`;
     await pipelinePage.enterPipelineName(pipelineName);
     await pipelinePage.savePipeline();
+    await ingestion(page);
+    // Verify the data ingested in destination and verify under logs page
+    await exploreStreamAndNavigateToPipeline(page, 'destination_node');
     await page.waitForTimeout(1000);
     await pipelinePage.searchPipeline(pipelineName);
     await page.waitForTimeout(1000);
@@ -616,208 +667,78 @@ test.describe("Pipeline testcases", () => {
     await page.getByText("Please connect all nodes").click();
   });
 
-  test("should upload an enrichment table under functions", async ({
+
+  test("should navigate to dashboard page if user accepts dialog box", async ({
     page,
   }) => {
     const pipelinePage = new PipelinePage(page);
-    await pipelinePage.navigateToAddEnrichmentTable();
-  
-    // Generate a unique file name and replace hyphens with underscores
-    let fileName = `enrichment_info_${uuidv4()}_csv`.replace(/-/g, "_");
-    console.log("Generated File Name:", fileName);
 
-    // Set the file to be uploaded
-    const fileContentPath = "../test-data/enrichment_info.csv";
-    const inputFile = await page.locator('input[type="file"]');
-    await inputFile.setInputFiles(fileContentPath);
-
-    // Enter the file name
-    await page.fill(".q-input > .q-field__inner > .q-field__control", fileName);
-
-    // Click on 'Save'
-    await page.getByText("Save").click({ force: true });
-
-    // Wait for the process to complete
-    await page.reload();
-    await page.waitForTimeout(3000);
-    await page.getByPlaceholder("Search Enrichment Table").fill(fileName);
-    await page.waitForTimeout(3000);
-    // Explore the uploaded table
-    await page.getByRole("button", { name: "Explore" }).click();
-    await page.locator('[data-test="log-table-column-0-_timestamp"]').click();
-    await page.locator('[data-test="close-dialog"]').click();
-    await page.waitForTimeout(3000);
-
-    // Return to the pipeline menu and verify the uploaded table
-    await page.locator('[data-test="menu-link-\\/pipeline-item"]').click();
-    await page.locator('[data-test="function-enrichment-table-tab"]').click();
-
-    // Verify and delete enrichment tables
-    await pipelinePage.deleteEnrichmentTableByName(fileName)
-  });
-
-  test.skip("should upload an enrichment table under functions with VRL", async ({
-    page,
-  }) => {
-    const pipelinePage = new PipelinePage(page);
-    await pipelinePage.navigateToAddEnrichmentTable();
-
-    // Generate a unique file name
-    let fileName = `protocols_${uuidv4()}_csv`;
-    console.log("Generated File Name:", fileName);
-    fileName = fileName.replace(/-/g, "_");
-
-    // Set the file to be uploaded
-    const fileContentPath = "../test-data/protocols.csv";
-    const inputFile = await page.locator('input[type="file"]');
-    await inputFile.setInputFiles(fileContentPath);
-
-    // Enter the file name
-    await page.fill(".q-input > .q-field__inner > .q-field__control", fileName);
-
-    // Click on 'Save'
-    await page.getByText("Save").click({ force: true });
-
-    // Reload the page to ensure table is updated
-    await page.reload();
-    await page.waitForTimeout(3000);
-
-    // Search for the uploaded file name
-    await page.getByPlaceholder("Search Enrichment Table").fill(fileName);
-    await page.waitForTimeout(3000);
-
-    // Verify the file name in the enrichment table
-    const rows = await page.locator("tbody tr");
-    let fileFound = false;
-
-    for (let i = 0; i < (await rows.count()); i++) {
-      const row = rows.nth(i);
-      const displayedName = await row
-        .locator("td.text-left")
-        .nth(1)
-        .textContent();
-      if (displayedName?.trim() === fileName) {
-        fileFound = true;
-        console.log("File found in table:", displayedName);
-        break;
-      }
-    }
-
-    if (!fileFound) {
-      throw new Error(`File name "${fileName}" not found in the table.`);
-    }
-
-    // Explore the file
-    await page.getByRole("button", { name: "Explore" }).click();
-    await page.locator("#fnEditor").getByLabel("Editor content;Press Alt+F1")
-      .fill(`
-abc, err = get_enrichment_table_record("${fileName}", {
-  "protocol_number": to_string!(.protocol_number)
-})
-.protocol_keyword = abc.keyword
-`);
-
-    // Wait briefly to allow input processing
+    await pipelinePage.openPipelineMenu();
     await page.waitForTimeout(1000);
+    await pipelinePage.addPipeline();
+    await pipelinePage.selectStream();
+    await pipelinePage.streamButton.waitFor({ state: "visible" });
+    await pipelinePage.dragStreamToTarget(pipelinePage.streamButton);
+    await pipelinePage.selectLogs();
 
-    // Apply the query
-    await applyQueryButton(page);
-
-    // Verify that no warning is shown for query execution
-    const warningElement = page.locator("text=warning Query execution");
-    await expect(warningElement).toBeHidden();
-
-    // Expand the first row in the logs table
-    await page.waitForTimeout(3000);
-    await page
-      .locator(
-        '[data-test="log-table-column-0-_timestamp"] [data-test="table-row-expand-menu"]'
-      )
-      .click();
-      await page.waitForTimeout(1000);
-    await page
-      .locator('[data-test="log-expand-detail-key-protocol_keyword-text"]')
-      .click();
-    // Return to the pipeline menu and delete the uploaded table
-    await page.locator('[data-test="menu-link-\\/pipeline-item"]').click();
-    await page.locator('[data-test="function-enrichment-table-tab"]').click();
-    await pipelinePage.deleteEnrichmentTableByName(fileName)
-
+    // Interact with stream name and save
+    await page.waitForTimeout(2000);
+    await pipelinePage.enterStreamName("e2e");
+    await pipelinePage.enterStreamName("e2e_automate");
+    await page.waitForTimeout(2000);
+    await pipelinePage.selectStreamOption();
+    await pipelinePage.saveInputNodeStream();
+    await page.locator('[data-test="menu-link-\\/dashboards-item"]').click();
+    page.once('dialog', async (dialog) => {
+      console.log(`Dialog message: ${dialog.message()}`);
+      await dialog.accept();
+    });
+    
+    // Click on the first menu link
+    await page.locator('[data-test="menu-link-\\/dashboards-item"]').click();
+    
+    // Assert the page navigates to the desired URL containing "dashboards"
+    await page.waitForURL(/dashboards/);
+    
   });
 
-
-  test("should display error when CSV not added in enrichment table", async ({
+  test("should stay on pipeline page if user dismisses the dialog box", async ({
     page,
   }) => {
     const pipelinePage = new PipelinePage(page);
-    await pipelinePage.navigateToAddEnrichmentTable();
-    await page.getByLabel('Name').fill('test')
-  
-    await page.getByRole('button', { name: 'Save' }).click();
-    await page.getByText('CSV File is required!').click();
-  })
 
+    await pipelinePage.openPipelineMenu();
+    await page.waitForTimeout(1000);
+    await pipelinePage.addPipeline();
+    await pipelinePage.selectStream();
+    await pipelinePage.streamButton.waitFor({ state: "visible" });
+    await pipelinePage.dragStreamToTarget(pipelinePage.streamButton);
+    await pipelinePage.selectLogs();
 
-  test("should append an enrichment table under functions", async ({ page }) => {
-    const pipelinePage = new PipelinePage(page);
-    await pipelinePage.navigateToAddEnrichmentTable();
-  
-    // Generate a unique file name and replace hyphens with underscores
-    let fileName = `append_${uuidv4()}_csv`.replace(/-/g, "_");
-    console.log("Generated File Name:", fileName);
-  
-    // First upload step
-    const fileContentPath = "../test-data/append.csv";
-    const inputFile = await page.locator('input[type="file"]');
-    await inputFile.setInputFiles(fileContentPath);
-  
-    // Enter the file name
-    await page.fill(".q-input > .q-field__inner > .q-field__control", fileName);
-  
-    // Click on 'Save'
-    await page.getByText("Save").click({ force: true });
-  
-    // Wait for the process to complete
-    await page.reload();
-    await page.waitForTimeout(3000);
-    await page.getByPlaceholder("Search Enrichment Table").fill(fileName);
-    await page.waitForTimeout(3000);
-  
-    // Explore the uploaded table
-    await page.getByRole("button", { name: "Explore" }).click();
-    await page.locator('[data-test="log-table-column-0-_timestamp"]').click();
-    await page.locator('[data-test="close-dialog"]').click();
-    await page.waitForTimeout(3000);
-  
-    // Navigate to append data to the existing enrichment table
-    await page.locator('[data-test="menu-link-\\/pipeline-item"]').click();
-    await page.locator('[data-test="function-enrichment-table-tab"]').click();
-    await page.getByPlaceholder("Search Enrichment Table").fill(fileName);
-    await page.getByRole("button", { name: "Enrichment Tables" }).click();
-  
-    // Select the append toggle
-    await page.getByLabel("Append data to existing").locator("div").first().click();
- 
-  
-    // Upload the CSV again for appending
-    await inputFile.setInputFiles(fileContentPath);
-    await page.waitForTimeout(3000);
-    await page.getByRole('button', { name: 'Save' }).click();
-    await page.waitForTimeout(3000);
-    await page.getByPlaceholder("Search Enrichment Table").fill(fileName);
-    await page.waitForTimeout(3000);
-  
-    // Explore the uploaded table
-    await page.getByRole("button", { name: "Explore" }).click();
+    // Interact with stream name and save
+    await page.waitForTimeout(2000);
+    await pipelinePage.enterStreamName("e2e");
+    await pipelinePage.enterStreamName("e2e_automate");
+    await page.waitForTimeout(2000);
+    await pipelinePage.selectStreamOption();
+    await pipelinePage.saveInputNodeStream();
+    await page.locator('[data-test="menu-link-\\/dashboards-item"]').click();
+    page.once('dialog', async (dialog) => {
+      console.log(`Dialog message: ${dialog.message()}`);
+      await dialog.dismiss();
+    });
+    
+    // Click on the first menu link
+    await page.locator('[data-test="menu-link-\\/dashboards-item"]').click();
+    
+    
+    // Assert page url to have pipeline
+    await expect(page).toHaveURL(/pipeline/);
 
-    const showingText = page.getByText('Showing 1 to 2 out of 2');
-  await expect(showingText).toBeVisible()
+    // Log confirmation
+    console.log('URL contains "pipeline"!')
+    
+  });
 
-  await page.locator('[data-test="menu-link-\\/pipeline-item"]').click();
-  await page.locator('[data-test="function-enrichment-table-tab"]').click();
-  await page.getByPlaceholder("Search Enrichment Table").fill(fileName);
-  await pipelinePage.deleteEnrichmentTableByName(fileName)
-  
-   });
   
 });
